requirements/local.txt
-r base.txt

black==24.3.0
blacken-docs==1.16.0
django-debug-toolbar==4.3.0
flake8==7.0.0
isort==5.13.2
pre-commit==3.7.0
